+++ src/test/java/lotto/domain/statisitc/MatchItemsTest.java
@@ -1,6 +1,7 @@
 package lotto.domain.statisitc;
 
 import lotto.domain.item.LottoTicket;
+import lotto.domain.item.MatchLotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.provider.Arguments;
